requirements.txt
pyaudio==0.2.13
numpy==1.21.5
sounddevice==0.4.5
openai-whisper==20230314
transformers==4.28.1
nltk==3.7
torch==1.13.1
torchaudio==0.13.1
scipy==1.10.1
PyQt5==5.15.6